binary-tree/traversal/preorder_test.go
package traversal

import (
	"fmt"
	"reflect"
	"testing"
)

func Test_preorderTraversal(t *testing.T) {
	type args struct {
		root *TreeNode
	}
	tests := []struct {
		name string
		args args
		want []int
	}{
		{
			name: "example 1",
			args: args{
				root: &TreeNode{
					Val: 1,
					Right: &TreeNode{
						Val: 2,
						Left: &TreeNode{
							Val: 3,
						},
					},
				},
			},
			want: []int{1, 2, 3},
		},
		{
			name: "test 1",
			args: args{
				root: &TreeNode{
					Val: 2,
					Left: &TreeNode{
						Val: 1,
					},
					Right: &TreeNode{
						Val: 4,
						Left: &TreeNode{
							Val: 3,
						},
						Right: &TreeNode{
							Val: 5,
						},
					},
				},
			},
			want: []int{2, 1, 4, 3, 5},
		},
		{
			name: "empty",
		},
	}

	for idx, f := range []func(*TreeNode) []int{
		preorderTraversal,
		preorderTraversal1,
		preorderTraversal2,
	} {
		t.Run(fmt.Sprintf("func#%d", idx), func(t *testing.T) {
			for _, tt := range tests {
				t.Run(tt.name, func(t *testing.T) {
					if got := f(tt.args.root); !reflect.DeepEqual(got, tt.want) {
						t.Errorf("preorderTraversal() = %v, want %v", got, tt.want)
					}
				})
			}
		})
	}
}
